refactor: wrap get function in getbasicinfo
